Refactor `PVSStudioParser` to remove the ErrorProne warnings.
