+++ src/main/java/edu/hm/hafner/analysis/registry/PvsStudioDescriptor.java
@@ -1,7 +1,7 @@
 package edu.hm.hafner.analysis.registry;
 
 import edu.hm.hafner.analysis.IssueParser;
-import edu.hm.hafner.analysis.parser.pvsstudio.PVSStudioParser;
+import edu.hm.hafner.analysis.parser.pvsstudio.PvsStudioParser;
 
 /**
  * A descriptor for the PVS-Studio static analyzer.
@@ -18,7 +18,7 @@ class PvsStudioDescriptor extends ParserDescriptor {
 
     @Override
     public IssueParser createParser(final Option... options) {
-        return new PVSStudioParser();
+        return new PvsStudioParser();
     }
 
     @Override
